rename RestJournalFøring.kt -> FerdigstillOppgaveKnyttJournalpostDto og flytt NavnOgIdent i samme fil
